+++ tests/test_cnn.py
@@ -10,7 +10,7 @@ from stable_baselines3 import A2C, DQN, PPO, SAC, TD3
 from stable_baselines3.common.envs import FakeImageEnv
 from stable_baselines3.common.preprocessing import is_image_space, is_image_space_channels_first
 from stable_baselines3.common.utils import zip_strict
-from stable_baselines3.common.vec_env import DummyVecEnv, VecFrameStack, VecTransposeImage, is_vecenv_wrapped
+from stable_baselines3.common.vec_env import DummyVecEnv, VecFrameStack, VecNormalize, VecTransposeImage, is_vecenv_wrapped
 
 
 @pytest.mark.parametrize("model_class", [A2C, PPO, SAC, TD3, DQN])
@@ -269,6 +269,10 @@ def test_image_space_checks():
     not_image_space = spaces.Box(0, 10, shape=(10, 10, 3), dtype=np.uint8)
     assert not is_image_space(not_image_space)
 
+    # Deactivate dtype and bound checking
+    normalized_image = spaces.Box(0, 1, shape=(10, 10, 3), dtype=np.float32)
+    assert is_image_space(normalized_image, normalized_image=True)
+
     # Not correct space
     not_image_space = spaces.Discrete(n=10)
     assert not is_image_space(not_image_space)
@@ -297,3 +301,44 @@ def test_image_space_checks():
     # Should raise a warning
     with pytest.warns(Warning):
         assert not is_image_space_channels_first(channel_mid_space)
+
+
+@pytest.mark.parametrize("model_class", [A2C, PPO, DQN, SAC, TD3])
+@pytest.mark.parametrize("normalize_images", [True, False])
+def test_image_like_input(model_class, normalize_images):
+    """
+    Check that we can handle image-like input (3D tensor)
+    when normalize_images=False
+    """
+    # Fake grayscale with frameskip
+    # Atari after preprocessing: 84x84x1, here we are using lower resolution
+    # to check that the network handle it automatically
+    env = FakeImageEnv(
+        screen_height=36,
+        screen_width=36,
+        n_channels=1,
+        channel_first=True,
+        discrete=model_class not in {SAC, TD3},
+    )
+    vec_env = VecNormalize(DummyVecEnv([lambda: env]))
+    # Reduce the size of the features
+    # deactivate normalization
+    kwargs = dict(
+        policy_kwargs=dict(
+            normalize_images=normalize_images,
+            features_extractor_kwargs=dict(features_dim=32),
+        ),
+        seed=1,
+    )
+
+    if model_class in {A2C, PPO}:
+        kwargs.update(dict(n_steps=64))
+    else:
+        # Avoid memory error when using replay buffer
+        # Reduce the size of the features
+        kwargs.update(dict(buffer_size=250))
+    if normalize_images:
+        with pytest.raises(AssertionError):
+            model_class("CnnPolicy", vec_env, **kwargs).learn(128)
+    else:
+        model_class("CnnPolicy", vec_env, **kwargs).learn(128)